--- docs/README.md
@@ -1,6 +1,0 @@
-# Introduction 
-This website contains data analysis and reports developed by Cal-ITP data analysts and scientists.
-
-## Source code
-
-All source code for these analyses and reports may be found [on GitHub](https://github.com/cal-itp/data-analyses).
