--- docs/_config.yml
@@ -1,41 +1,0 @@
-# Book settings
-# Learn more at https://jupyterbook.org/customize/config.html
-
-title: Cal-ITP Analytics Portfolio
-author: Cal-ITP
-#logo: calitp_logo_MAIN.png
-
-# Force re-execution of notebooks on each build.
-# See https://jupyterbook.org/content/execute.html
-execute:
-  execute_notebooks: 'off'
-  allow_errors: false
-  timeout: -1
-
-# Define the name of the latex output file for PDF builds
-latex:
-  latex_documents:
-    targetname: book.tex
-
-launch_buttons:
-  binderhub_url: "https://mybinder.org"
-  jupyterhub_url: "https://hubtest.k8s.calitp.jarv.us"
-  thebe: true
-
-repository:
-  url: https://github.com/cal-itp/data-analyses/  # Online location of your book
-  path_to_book: docs  # Optional path to your book, relative to the repository root
-  branch: main  # Which branch of the repository should be used when creating links (optional)
-
-# Add GitHub buttons to your book
-# See https://jupyterbook.org/customize/config.html#add-a-link-to-your-repository
-html:
-  use_issues_button: true
-  use_repository_button: true
-  use_edit_page_button: true
-  google_analytics_id: G-SZB618VNBZ
-
-sphinx:
-  config:
-    html_js_files:
-    - https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js
