--- docs/_toc.yml
@@ -1,9 +1,0 @@
-
-format: jb-book
-root: README
-parts:
-  - #caption: DLA Analyses by District
-    chapters:
-      - file: dla/dla_analysis
-        sections:
-          - glob: dla/e76_obligated_funds/*
