+++ portfolio/_toc.yml
@@ -1,0 +1,8 @@
+
+format: jb-book
+root: README
+parts:
+  - chapters:
+      - file: dla/dla_analysis
+        sections:
+          - glob: dla/e76_obligated_funds/*
