+++ portfolio/dla2/dla_analysis.md
@@ -1,0 +1,6 @@
+# DLA Grant Analysis by District
+
+Analysis of grants provided through the Division of Local Assistance.
+
+This is currently an example of static, hosted reports built on top of
+parameterized Jupyter notebooks.
