+++ portfolio/index/_redirects
@@ -1,0 +1,2 @@
+/dla/* https://dla-cal-itp-data-analyses.netlify.app//:splat  200
+/dla2/* https://dla2-cal-itp-data-analyses.netlify.app//:splat  200
