+++ portfolio/templates/_redirects
@@ -1,0 +1,3 @@
+{% for analysis in analyses -%}
+/{{ analysis }}/* https://{{ analysis }}-cal-itp-data-analyses.netlify.app//:splat  200
+{% endfor %}
