--- reports_config.yml
@@ -1,6 +1,0 @@
-dla:
-  notebook: ./dla/e76_obligated_funds/prototype-district.ipynb
-  params:
-    district:
-      - 4
-      - 7
