--- requirements.txt
@@ -1,8 +1,0 @@
-nbformat==5.1.3
-papermill==2.3.4
-pydantic==1.9.0
-pyyaml==6.0
-typer==0.4.1
-jupyter-book==0.12.2
-intake==0.6.4
-humanize==4.0.0
